% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models-BaselineHazard.R
\name{BaselineHazard}
\alias{BaselineHazard}
\title{Hazard function (R6 class)}
\description{
Hazard function (R6 class)

Hazard function (R6 class)
}
\section{Super class}{
\code{\link[sfgp:StanCodeCreator]{sfgp::StanCodeCreator}} -> \code{BaselineHazard}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BaselineHazard-stanname_t}{\code{BaselineHazard$stanname_t()}}
\item \href{#method-BaselineHazard-clone}{\code{BaselineHazard$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="sfgp" data-topic="StanCodeCreator" data-id="print"><a href='../../sfgp/html/StanCodeCreator.html#method-StanCodeCreator-print'><code>sfgp::StanCodeCreator$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sfgp" data-topic="StanCodeCreator" data-id="stancode_data"><a href='../../sfgp/html/StanCodeCreator.html#method-StanCodeCreator-stancode_data'><code>sfgp::StanCodeCreator$stancode_data()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sfgp" data-topic="StanCodeCreator" data-id="stancode_functions"><a href='../../sfgp/html/StanCodeCreator.html#method-StanCodeCreator-stancode_functions'><code>sfgp::StanCodeCreator$stancode_functions()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sfgp" data-topic="StanCodeCreator" data-id="stancode_gq"><a href='../../sfgp/html/StanCodeCreator.html#method-StanCodeCreator-stancode_gq'><code>sfgp::StanCodeCreator$stancode_gq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sfgp" data-topic="StanCodeCreator" data-id="stancode_model"><a href='../../sfgp/html/StanCodeCreator.html#method-StanCodeCreator-stancode_model'><code>sfgp::StanCodeCreator$stancode_model()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sfgp" data-topic="StanCodeCreator" data-id="stancode_pars"><a href='../../sfgp/html/StanCodeCreator.html#method-StanCodeCreator-stancode_pars'><code>sfgp::StanCodeCreator$stancode_pars()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sfgp" data-topic="StanCodeCreator" data-id="stancode_tdata"><a href='../../sfgp/html/StanCodeCreator.html#method-StanCodeCreator-stancode_tdata'><code>sfgp::StanCodeCreator$stancode_tdata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sfgp" data-topic="StanCodeCreator" data-id="stancode_tpars"><a href='../../sfgp/html/StanCodeCreator.html#method-StanCodeCreator-stancode_tpars'><code>sfgp::StanCodeCreator$stancode_tpars()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sfgp" data-topic="StanCodeCreator" data-id="stanfiles_functions"><a href='../../sfgp/html/StanCodeCreator.html#method-StanCodeCreator-stanfiles_functions'><code>sfgp::StanCodeCreator$stanfiles_functions()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sfgp" data-topic="StanCodeCreator" data-id="string"><a href='../../sfgp/html/StanCodeCreator.html#method-StanCodeCreator-string'><code>sfgp::StanCodeCreator$string()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaselineHazard-stanname_t"></a>}}
\if{latex}{\out{\hypertarget{method-BaselineHazard-stanname_t}{}}}
\subsection{Method \code{stanname_t()}}{
Get name of the hazard time variable in 'Stan' code
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaselineHazard$stanname_t(datanames)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{datanames}}{names of data sets}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaselineHazard-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BaselineHazard-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaselineHazard$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
