% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-treatment_effect.R
\name{dor_diff}
\alias{dor_diff}
\title{Get duration differences to control}
\usage{
dor_diff(dor, control_group_name = NULL, gt = FALSE)
}
\arguments{
\item{dor}{Vector of depth \code{rvar}s.}

\item{control_group_name}{Control arm name.}

\item{gt}{See the \code{control_group_name} argument.}
}
\description{
Get duration differences to control
}
