% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-treatment_effect.R
\name{dur_diff}
\alias{dur_diff}
\title{Get duration differences to control}
\usage{
dur_diff(dur, control_group_name = NULL, gt = FALSE)
}
\arguments{
\item{dur}{List of duration \code{rvar} vectors.}

\item{control_group_name}{Control arm name.}

\item{gt}{See the \code{control_group_name} argument.}
}
\value{
\itemize{
\item A list with \code{rvar}s that are differences to control, or
\item if \code{gt = TRUE}, matrix with number of rows equal to number of
arms, where elements are for the proportion of samples larger than control
(out of non-equal samples)
}
}
\description{
Get duration differences to control
}
