% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data_jm}
\alias{example_data_jm}
\title{Simulated joint longitudinal and time-to-event test data}
\format{
A list of two data frames \code{lon} (longitudinal data) and
\code{tte} (time-to-event data). Generated using \code{simjm}.
}
\usage{
example_data_jm
}
\description{
Simulated joint longitudinal and time-to-event test data
}
\seealso{
Other built-in datasets: 
\code{\link{testdata}}
}
\concept{built-in datasets}
\keyword{datasets}
