% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{expose_stan_functions}
\alias{expose_stan_functions}
\title{Expose functions in a given Stan file to R}
\usage{
expose_stan_functions(stan_file, ...)
}
\arguments{
\item{stan_file}{Path to the Stan file relative to the directory
\code{inst/functions}, without the \code{.stan} suffix.}

\item{...}{Optional arguments to \code{cmdstanr::cmdstan_model()}.}
}
\description{
Does not seem to work.
}
