% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models-JointModel.R
\name{JointModel}
\alias{JointModel}
\title{The joint model class (R6 class)}
\description{
The joint model class (R6 class)

The joint model class (R6 class)
}
\section{Super classes}{
\code{\link[sfgp:StanCodeCreator]{sfgp::StanCodeCreator}} -> \code{\link[sfgp:StanModel]{sfgp::StanModel}} -> \code{JointModel}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{tte}}{The time-to-event model, has class \code{\link{TTEModel}}.}

\item{\code{lon}}{The longitudinal model, has class \code{\link{TSModel}}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-JointModel-new}{\code{JointModel$new()}}
\item \href{#method-JointModel-string}{\code{JointModel$string()}}
\item \href{#method-JointModel-create_standata}{\code{JointModel$create_standata()}}
\item \href{#method-JointModel-fit}{\code{JointModel$fit()}}
\item \href{#method-JointModel-clone}{\code{JointModel$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="sfgp" data-topic="StanCodeCreator" data-id="print"><a href='../../sfgp/html/StanCodeCreator.html#method-StanCodeCreator-print'><code>sfgp::StanCodeCreator$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sfgp" data-topic="StanCodeCreator" data-id="stancode_data"><a href='../../sfgp/html/StanCodeCreator.html#method-StanCodeCreator-stancode_data'><code>sfgp::StanCodeCreator$stancode_data()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sfgp" data-topic="StanCodeCreator" data-id="stancode_functions"><a href='../../sfgp/html/StanCodeCreator.html#method-StanCodeCreator-stancode_functions'><code>sfgp::StanCodeCreator$stancode_functions()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sfgp" data-topic="StanCodeCreator" data-id="stancode_gq"><a href='../../sfgp/html/StanCodeCreator.html#method-StanCodeCreator-stancode_gq'><code>sfgp::StanCodeCreator$stancode_gq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sfgp" data-topic="StanCodeCreator" data-id="stancode_model"><a href='../../sfgp/html/StanCodeCreator.html#method-StanCodeCreator-stancode_model'><code>sfgp::StanCodeCreator$stancode_model()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sfgp" data-topic="StanCodeCreator" data-id="stancode_pars"><a href='../../sfgp/html/StanCodeCreator.html#method-StanCodeCreator-stancode_pars'><code>sfgp::StanCodeCreator$stancode_pars()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sfgp" data-topic="StanCodeCreator" data-id="stancode_tdata"><a href='../../sfgp/html/StanCodeCreator.html#method-StanCodeCreator-stancode_tdata'><code>sfgp::StanCodeCreator$stancode_tdata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sfgp" data-topic="StanCodeCreator" data-id="stancode_tpars"><a href='../../sfgp/html/StanCodeCreator.html#method-StanCodeCreator-stancode_tpars'><code>sfgp::StanCodeCreator$stancode_tpars()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sfgp" data-topic="StanCodeCreator" data-id="stanfiles_functions"><a href='../../sfgp/html/StanCodeCreator.html#method-StanCodeCreator-stanfiles_functions'><code>sfgp::StanCodeCreator$stanfiles_functions()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sfgp" data-topic="StanModel" data-id="compile"><a href='../../sfgp/html/StanModel.html#method-StanModel-compile'><code>sfgp::StanModel$compile()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sfgp" data-topic="StanModel" data-id="create_stancode"><a href='../../sfgp/html/StanModel.html#method-StanModel-create_stancode'><code>sfgp::StanModel$create_stancode()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sfgp" data-topic="StanModel" data-id="create_stanmodel"><a href='../../sfgp/html/StanModel.html#method-StanModel-create_stanmodel'><code>sfgp::StanModel$create_stanmodel()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sfgp" data-topic="StanModel" data-id="get_stanmodel"><a href='../../sfgp/html/StanModel.html#method-StanModel-get_stanmodel'><code>sfgp::StanModel$get_stanmodel()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JointModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-JointModel-new}{}}}
\subsection{Method \code{new()}}{
Create model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JointModel$new(lon, h0 = NULL, compile = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lon}}{An object of class \code{\link{StanModel}}.}

\item{\code{h0}}{An object of class \code{\link{BaselineHazard}}.}

\item{\code{compile}}{Should the 'Stan' model code be created and compiled.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JointModel-string"></a>}}
\if{latex}{\out{\hypertarget{method-JointModel-string}{}}}
\subsection{Method \code{string()}}{
The model description as a string
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JointModel$string()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JointModel-create_standata"></a>}}
\if{latex}{\out{\hypertarget{method-JointModel-create_standata}{}}}
\subsection{Method \code{create_standata()}}{
Create the 'Stan' data list from data frames.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JointModel$create_standata(
  data_lon,
  data_tte,
  data_grid = NULL,
  term_confs = NULL,
  num_bf = NULL,
  scale_bf = NULL,
  skip_transform = NULL,
  prior_only = FALSE,
  set_transforms = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_lon}}{Longitudinal data, a data frame.}

\item{\code{data_tte}}{The events data, a data frame.}

\item{\code{data_grid}}{The integration grid data, a data frame.}

\item{\code{term_confs}}{A list that specifies configuration of model terms.
If name of any term is not found from the list, \code{$default_conf()}
of that \code{FormulaTerm} is used.}

\item{\code{num_bf}}{If not \code{NULL}, configurations of all
\code{GPTerm}s are updated with this value.}

\item{\code{scale_bf}}{If not \code{NULL}, configurations of all
\code{GPTerm}s are updated with this value.}

\item{\code{skip_transform}}{Term names whose input transform should be
skipped.}

\item{\code{prior_only}}{Sample from prior only?}

\item{\code{set_transforms}}{If longitudinal data transforms should be set
based on the given \code{data_lon}. This should be \code{TRUE} when
fitting a model, and \code{FALSE} when computing predictions using GQ.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JointModel-fit"></a>}}
\if{latex}{\out{\hypertarget{method-JointModel-fit}{}}}
\subsection{Method \code{fit()}}{
Fit the model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JointModel$fit(
  data_lon,
  data_tte,
  data_grid,
  term_confs = NULL,
  num_bf = NULL,
  scale_bf = NULL,
  skip_transform = NULL,
  prior_only = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_lon}}{Longitudinal data, a data frame.}

\item{\code{data_tte}}{The events data, a data frame.}

\item{\code{data_grid}}{The integration grid data, a data frame. Can be
created using \code{\link{create_jm_grid}}.}

\item{\code{term_confs}}{A list that specifies configuration of model terms.
If name of any term is not found from the list, \code{$default_conf()}
of that \code{FormulaTerm} is used.}

\item{\code{num_bf}}{If not \code{NULL}, configurations of all
\code{GPTerm}s are updated with this value.}

\item{\code{scale_bf}}{If not \code{NULL}, configurations of all
\code{GPTerm}s are updated with this value.}

\item{\code{skip_transform}}{Term names whose input transform should be
skipped.}

\item{\code{prior_only}}{Sample from prior only.}

\item{\code{...}}{Arguments passed to \code{sample} method of the
'CmdStanR' model.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{\link{JointModelFit}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JointModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-JointModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JointModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
