% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fits-JointModelFit.R
\name{JointModelFit}
\alias{JointModelFit}
\title{The time-to-event model fit class}
\description{
The time-to-event model fit class

The time-to-event model fit class
}
\section{Super classes}{
\code{\link[sfgp:StanModelFit]{sfgp::StanModelFit}} -> \code{\link[sfgp:TSModelFit]{sfgp::TSModelFit}} -> \code{JointModelFit}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-JointModelFit-print}{\code{JointModelFit$print()}}
\item \href{#method-JointModelFit-get_model}{\code{JointModelFit$get_model()}}
\item \href{#method-JointModelFit-function_draws_tte}{\code{JointModelFit$function_draws_tte()}}
\item \href{#method-JointModelFit-h0}{\code{JointModelFit$h0()}}
\item \href{#method-JointModelFit-instant_hazard}{\code{JointModelFit$instant_hazard()}}
\item \href{#method-JointModelFit-cumulative_hazard}{\code{JointModelFit$cumulative_hazard()}}
\item \href{#method-JointModelFit-survival_function}{\code{JointModelFit$survival_function()}}
\item \href{#method-JointModelFit-plot_hazard}{\code{JointModelFit$plot_hazard()}}
\item \href{#method-JointModelFit-predict}{\code{JointModelFit$predict()}}
\item \href{#method-JointModelFit-predict_time}{\code{JointModelFit$predict_time()}}
\item \href{#method-JointModelFit-clone}{\code{JointModelFit$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="sfgp" data-topic="StanModelFit" data-id="datanames"><a href='../../sfgp/html/StanModelFit.html#method-StanModelFit-datanames'><code>sfgp::StanModelFit$datanames()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sfgp" data-topic="StanModelFit" data-id="draws"><a href='../../sfgp/html/StanModelFit.html#method-StanModelFit-draws'><code>sfgp::StanModelFit$draws()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sfgp" data-topic="StanModelFit" data-id="get_data"><a href='../../sfgp/html/StanModelFit.html#method-StanModelFit-get_data'><code>sfgp::StanModelFit$get_data()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sfgp" data-topic="StanModelFit" data-id="get_stan_data"><a href='../../sfgp/html/StanModelFit.html#method-StanModelFit-get_stan_data'><code>sfgp::StanModelFit$get_stan_data()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sfgp" data-topic="StanModelFit" data-id="get_stan_fit"><a href='../../sfgp/html/StanModelFit.html#method-StanModelFit-get_stan_fit'><code>sfgp::StanModelFit$get_stan_fit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sfgp" data-topic="StanModelFit" data-id="gq"><a href='../../sfgp/html/StanModelFit.html#method-StanModelFit-gq'><code>sfgp::StanModelFit$gq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sfgp" data-topic="StanModelFit" data-id="log_z_pars"><a href='../../sfgp/html/StanModelFit.html#method-StanModelFit-log_z_pars'><code>sfgp::StanModelFit$log_z_pars()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sfgp" data-topic="StanModelFit" data-id="loglik"><a href='../../sfgp/html/StanModelFit.html#method-StanModelFit-loglik'><code>sfgp::StanModelFit$loglik()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sfgp" data-topic="TSModelFit" data-id="create_functiondraws"><a href='../../sfgp/html/TSModelFit.html#method-TSModelFit-create_functiondraws'><code>sfgp::TSModelFit$create_functiondraws()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sfgp" data-topic="TSModelFit" data-id="fit_quality_df"><a href='../../sfgp/html/TSModelFit.html#method-TSModelFit-fit_quality_df'><code>sfgp::TSModelFit$fit_quality_df()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sfgp" data-topic="TSModelFit" data-id="fit_quality_summary"><a href='../../sfgp/html/TSModelFit.html#method-TSModelFit-fit_quality_summary'><code>sfgp::TSModelFit$fit_quality_summary()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sfgp" data-topic="TSModelFit" data-id="function_draws"><a href='../../sfgp/html/TSModelFit.html#method-TSModelFit-function_draws'><code>sfgp::TSModelFit$function_draws()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sfgp" data-topic="TSModelFit" data-id="initialize"><a href='../../sfgp/html/TSModelFit.html#method-TSModelFit-initialize'><code>sfgp::TSModelFit$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sfgp" data-topic="TSModelFit" data-id="measurement_error"><a href='../../sfgp/html/TSModelFit.html#method-TSModelFit-measurement_error'><code>sfgp::TSModelFit$measurement_error()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sfgp" data-topic="TSModelFit" data-id="plot"><a href='../../sfgp/html/TSModelFit.html#method-TSModelFit-plot'><code>sfgp::TSModelFit$plot()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JointModelFit-print"></a>}}
\if{latex}{\out{\hypertarget{method-JointModelFit-print}{}}}
\subsection{Method \code{print()}}{
Print object description.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JointModelFit$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JointModelFit-get_model"></a>}}
\if{latex}{\out{\hypertarget{method-JointModelFit-get_model}{}}}
\subsection{Method \code{get_model()}}{
Get the entire joint model or a particular submodel.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JointModelFit$get_model(name = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of the submodel. If \code{NULL} (default), the entire
joint model object is returned.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JointModelFit-function_draws_tte"></a>}}
\if{latex}{\out{\hypertarget{method-JointModelFit-function_draws_tte}{}}}
\subsection{Method \code{function_draws_tte()}}{
Extract one TTE-related function as a \code{\link{FunctionDraws}} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JointModelFit$function_draws_tte(component = "inst_haz", dataname = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{component}}{Base name of the function in the 'Stan' code.}

\item{\code{dataname}}{Data where the function was evaluated.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JointModelFit-h0"></a>}}
\if{latex}{\out{\hypertarget{method-JointModelFit-h0}{}}}
\subsection{Method \code{h0()}}{
Extract baseline hazard as a \code{\link{FunctionDraws}} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JointModelFit$h0(dataname = "GRD")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataname}}{Data where the function was evaluated.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JointModelFit-instant_hazard"></a>}}
\if{latex}{\out{\hypertarget{method-JointModelFit-instant_hazard}{}}}
\subsection{Method \code{instant_hazard()}}{
Extract instant hazard as a \code{\link{FunctionDraws}} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JointModelFit$instant_hazard(dataname = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataname}}{Data where the function was evaluated.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JointModelFit-cumulative_hazard"></a>}}
\if{latex}{\out{\hypertarget{method-JointModelFit-cumulative_hazard}{}}}
\subsection{Method \code{cumulative_hazard()}}{
Extract cumulative hazard as a \code{\link{FunctionDraws}} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JointModelFit$cumulative_hazard(dataname = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataname}}{Data where the function was evaluated.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JointModelFit-survival_function"></a>}}
\if{latex}{\out{\hypertarget{method-JointModelFit-survival_function}{}}}
\subsection{Method \code{survival_function()}}{
Extract survival function as a \code{\link{FunctionDraws}} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JointModelFit$survival_function(dataname = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataname}}{Data where the function was evaluated.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JointModelFit-plot_hazard"></a>}}
\if{latex}{\out{\hypertarget{method-JointModelFit-plot_hazard}{}}}
\subsection{Method \code{plot_hazard()}}{
Visualize model fit (instant hazard).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JointModelFit$plot_hazard(plot_events = TRUE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{plot_events}}{Should the event occurrences be plotted?
Compute predictions at test points}

\item{\code{...}}{Arguments passed to the plot method of
\code{\link{FunctionDraws}}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JointModelFit-predict"></a>}}
\if{latex}{\out{\hypertarget{method-JointModelFit-predict}{}}}
\subsection{Method \code{predict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JointModelFit$predict(
  data_lon = NULL,
  data_tte = NULL,
  data_grid = NULL,
  fitted_params = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_lon}}{Data frame. If \code{NULL}, the data used in fitting
is used.}

\item{\code{data_tte}}{Data frame. If \code{NULL}, the data used in fitting
is used.}

\item{\code{data_grid}}{Data frame. If \code{NULL}, the data used in fitting
is used.}

\item{\code{fitted_params}}{Parameters or model fit.}

\item{\code{...}}{Other arguments to \code{generate_quantities()}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{\link{JointModelFit}} object.
Compute predictions at given time points
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JointModelFit-predict_time"></a>}}
\if{latex}{\out{\hypertarget{method-JointModelFit-predict_time}{}}}
\subsection{Method \code{predict_time()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JointModelFit$predict_time(
  t_test,
  t_var = "t",
  y_test = NULL,
  t_var_copy = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{t_test}}{The test time points}

\item{\code{t_var}}{Name of the time variable}

\item{\code{y_test}}{Test y}

\item{\code{t_var_copy}}{Names of other continuous covariates to which
\code{t_test} should be copied.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new fit object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JointModelFit-clone"></a>}}
\if{latex}{\out{\hypertarget{method-JointModelFit-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JointModelFit$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
