% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-treatment_effect.R
\name{plot_dor}
\alias{plot_dor}
\title{Visualize depth of response}
\usage{
plot_dor(te, halfeye = FALSE, control_group_name = NULL, ...)
}
\arguments{
\item{te}{A list returned by \code{\link{treatment_effect}}.}

\item{halfeye}{use halfeye plot instead of pointinterval}

\item{control_group_name}{If given, the values are differences to control,
or proportion of samples larger than control (out of non-equal samples) if
\code{gt = TRUE}.}

\item{...}{arguments passed to ggdist}
}
\description{
Visualize depth of response
}
