% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-treatment_effect.R
\name{plot_dur}
\alias{plot_dur}
\title{Visualize duration of response}
\usage{
plot_dur(
  te,
  only_responding = FALSE,
  metric = "duration",
  halfeye = FALSE,
  control_group_name = NULL,
  ...
)
}
\arguments{
\item{te}{A list returned by \code{\link{treatment_effect}}.}

\item{only_responding}{Compute only for the trajectories that have response.
Currently not implemented!}

\item{metric}{Can be one of \code{"duration"}, \code{"start"}, \code{"end"},
or \code{"TTB120"}.}

\item{halfeye}{use halfeye plot instead of pointinterval}

\item{control_group_name}{If given, the values are differences to control,
or proportion of samples larger than control (out of non-equal samples) if
\code{gt = TRUE}.}

\item{...}{arguments passed to \code{ggdist}.}
}
\description{
Visualize duration of response
}
