% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-treatment_effect.R
\name{plot_metric}
\alias{plot_metric}
\title{Example implementation of plotting trajectory metrics.}
\usage{
plot_metric(df, metric = "depth", group_var = "arm")
}
\arguments{
\item{df}{A full long data frame returned by
\code{\link{trajectory_metrics}}.}

\item{metric}{Name of the metric. See possible metric names in
documentation of \code{\link{trajectory_metrics}}.}

\item{group_var}{Name of grouping variable.}
}
\description{
Example implementation of plotting trajectory metrics.
}
