% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{plot_rvar_df}
\alias{plot_rvar_df}
\title{Visualize 'rvar' in data frame}
\usage{
plot_rvar_df(
  df,
  rvar_name,
  group_var,
  horizontal = FALSE,
  halfeye = FALSE,
  ...
)
}
\arguments{
\item{df}{the data frame}

\item{rvar_name}{name of the 'rvar' column}

\item{group_var}{name of the grouping variable}

\item{horizontal}{orient horizontally?}

\item{halfeye}{use halfeye plot instead of pointinterval}

\item{...}{arguments passed to ggdist}
}
\description{
Visualize 'rvar' in data frame
}
