% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-treatment_effect.R
\name{predict_new_subjects}
\alias{predict_new_subjects}
\title{Generate predictions for new imaginary subjects from each group}
\usage{
predict_new_subjects(
  fit_post,
  fit_prior,
  time_var,
  t_pred = NULL,
  group_var = "arm",
  zero_log_z = FALSE,
  prior_param_names = NULL
)
}
\arguments{
\item{fit_post}{posterior fit}

\item{fit_prior}{prior fit}

\item{time_var}{time variable}

\item{t_pred}{Vector of time points. Chosen automatically if \code{NULL}.}

\item{group_var}{grouping variable}

\item{zero_log_z}{Use zero for all \code{log_z_*} parameters. This
effectively takes group means of all \code{HierOffsetTerm}s.}

\item{prior_param_names}{Names of parameters for which prior draws should
be used. If \code{NULL} (default), these are attempted be to detected
automatically. These are usually all individual-specific parameters whose
values we do not know for "new" subjects so we draw them from the prior.}
}
\description{
Generate predictions for new imaginary subjects from each group
}
