% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models-TSModel.R
\name{prior_sigma_informed}
\alias{prior_sigma_informed}
\title{Create a prior of noise parameter based on information about expected
amount of measurement error}
\usage{
prior_sigma_informed(p = 0.065, p_sd = 0.015)
}
\arguments{
\item{p}{A priori the most likely value of measurement error percentage}

\item{p_sd}{Prior std in measurement error percentage}
}
\description{
Create a prior of noise parameter based on information about expected
amount of measurement error
}
