% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-treatment_effect.R
\name{sample_subjects}
\alias{sample_subjects}
\title{Draw one subjects from each group}
\usage{
sample_subjects(data, id_var = "id", group_var = "arm")
}
\arguments{
\item{data}{original data frame}

\item{id_var}{name of the subject identifier variable}

\item{group_var}{grouping variable}
}
\value{
new data frame
}
\description{
Draw one subjects from each group
}
