% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-treatment_effect.R
\name{summarize_dur}
\alias{summarize_dur}
\title{Summarize duration of response}
\usage{
summarize_dur(te, only_responding = FALSE)
}
\arguments{
\item{te}{A list returned by \code{\link{treatment_effect}}.}

\item{only_responding}{Compute only for the trajectories that have response.
Currently not implemented!}
}
\description{
Summarize duration of response
}
