% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{testdata}
\alias{testdata}
\title{A small artificial test data}
\format{
A data frame with 176 rows and 6 variables:
\describe{
  \item{id}{factor, individual id}
  \item{time}{numeric, a continuous variable}
  \item{arm}{factor, three levels}
  \item{sex}{factor, two levels}
  \item{weight}{numeric, a continuous variable}
  \item{y}{numeric, a continuous variable}
  \item{f_true}{numeric, the true signal used to generate the data}
}
}
\usage{
testdata
}
\description{
A small artificial test data
}
\concept{built-in datasets}
\keyword{datasets}
