% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-treatment_effect.R
\name{trajectory_metrics}
\alias{trajectory_metrics}
\title{Compute trajectory metrics.}
\usage{
trajectory_metrics(trajectories, id_var, time_var)
}
\arguments{
\item{trajectories}{An object of class \code{\link{FunctionDraws}}.}

\item{id_var}{Name of id variable.}

\item{time_var}{Name of the time variable.}
}
\value{
The following metrics are computed for each draw for each id.
\itemize{
 \item \code{depth} = depth of response (RECIST)
 \item \code{depth_br} = depth of response (best response definition, can
 be negative)
 \item \code{duration} = duration of response (RECIST)
 \item \code{start} = start time of response (RECIST)
 \item \code{end} = end time of response (RECIST)
 \item \code{exists} = whether response exists (RECIST)
 \item \code{endures} = whether response endures whole time interval (RECIST)
 \item \code{ttb120} = Time to 1.2 x baseline (different duration metric).
}
See the \code{treatment-effect} vignette for definition of the metrics.
}
\description{
Compute trajectory metrics.
}
