% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models-TTEModel.R
\name{TTEModel}
\alias{TTEModel}
\title{The time-to-event model class (R6 class)}
\description{
The time-to-event model class (R6 class)

The time-to-event model class (R6 class)
}
\section{Super classes}{
\code{\link[sfgp:StanCodeCreator]{sfgp::StanCodeCreator}} -> \code{\link[sfgp:StanModel]{sfgp::StanModel}} -> \code{TTEModel}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{h0}}{An object of class \code{\link{BaselineHazard}}.}

\item{\code{id_var}}{name of id variable}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TTEModel-stanname_id}{\code{TTEModel$stanname_id()}}
\item \href{#method-TTEModel-new}{\code{TTEModel$new()}}
\item \href{#method-TTEModel-create_standata}{\code{TTEModel$create_standata()}}
\item \href{#method-TTEModel-clone}{\code{TTEModel$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="sfgp" data-topic="StanCodeCreator" data-id="print"><a href='../../sfgp/html/StanCodeCreator.html#method-StanCodeCreator-print'><code>sfgp::StanCodeCreator$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sfgp" data-topic="StanCodeCreator" data-id="stancode_data"><a href='../../sfgp/html/StanCodeCreator.html#method-StanCodeCreator-stancode_data'><code>sfgp::StanCodeCreator$stancode_data()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sfgp" data-topic="StanCodeCreator" data-id="stancode_functions"><a href='../../sfgp/html/StanCodeCreator.html#method-StanCodeCreator-stancode_functions'><code>sfgp::StanCodeCreator$stancode_functions()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sfgp" data-topic="StanCodeCreator" data-id="stancode_gq"><a href='../../sfgp/html/StanCodeCreator.html#method-StanCodeCreator-stancode_gq'><code>sfgp::StanCodeCreator$stancode_gq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sfgp" data-topic="StanCodeCreator" data-id="stancode_model"><a href='../../sfgp/html/StanCodeCreator.html#method-StanCodeCreator-stancode_model'><code>sfgp::StanCodeCreator$stancode_model()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sfgp" data-topic="StanCodeCreator" data-id="stancode_pars"><a href='../../sfgp/html/StanCodeCreator.html#method-StanCodeCreator-stancode_pars'><code>sfgp::StanCodeCreator$stancode_pars()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sfgp" data-topic="StanCodeCreator" data-id="stancode_tdata"><a href='../../sfgp/html/StanCodeCreator.html#method-StanCodeCreator-stancode_tdata'><code>sfgp::StanCodeCreator$stancode_tdata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sfgp" data-topic="StanCodeCreator" data-id="stancode_tpars"><a href='../../sfgp/html/StanCodeCreator.html#method-StanCodeCreator-stancode_tpars'><code>sfgp::StanCodeCreator$stancode_tpars()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sfgp" data-topic="StanCodeCreator" data-id="stanfiles_functions"><a href='../../sfgp/html/StanCodeCreator.html#method-StanCodeCreator-stanfiles_functions'><code>sfgp::StanCodeCreator$stanfiles_functions()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sfgp" data-topic="StanCodeCreator" data-id="string"><a href='../../sfgp/html/StanCodeCreator.html#method-StanCodeCreator-string'><code>sfgp::StanCodeCreator$string()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sfgp" data-topic="StanModel" data-id="compile"><a href='../../sfgp/html/StanModel.html#method-StanModel-compile'><code>sfgp::StanModel$compile()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sfgp" data-topic="StanModel" data-id="create_stancode"><a href='../../sfgp/html/StanModel.html#method-StanModel-create_stancode'><code>sfgp::StanModel$create_stancode()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sfgp" data-topic="StanModel" data-id="create_stanmodel"><a href='../../sfgp/html/StanModel.html#method-StanModel-create_stanmodel'><code>sfgp::StanModel$create_stanmodel()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sfgp" data-topic="StanModel" data-id="get_stanmodel"><a href='../../sfgp/html/StanModel.html#method-StanModel-get_stanmodel'><code>sfgp::StanModel$get_stanmodel()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TTEModel-stanname_id"></a>}}
\if{latex}{\out{\hypertarget{method-TTEModel-stanname_id}{}}}
\subsection{Method \code{stanname_id()}}{
Get name of id variable in 'Stan' code.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TTEModel$stanname_id(datanames)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{datanames}}{names of data sets}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TTEModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-TTEModel-new}{}}}
\subsection{Method \code{new()}}{
Create model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TTEModel$new(h0 = NULL, link_name = "f_sum")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{h0}}{An object of class \code{\link{BaselineHazard}}.}

\item{\code{link_name}}{Base name of the link variable in 'Stan' code.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TTEModel-create_standata"></a>}}
\if{latex}{\out{\hypertarget{method-TTEModel-create_standata}{}}}
\subsection{Method \code{create_standata()}}{
Create the 'Stan' data list from a data vector.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TTEModel$create_standata(event)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{event}}{a logical vector}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TTEModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TTEModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TTEModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
