% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models-BaselineHazard.R
\name{WeibullHazard}
\alias{WeibullHazard}
\title{Weibull hazard function}
\description{
Weibull hazard function

Weibull hazard function
}
\section{Super classes}{
\code{\link[sfgp:StanCodeCreator]{sfgp::StanCodeCreator}} -> \code{\link[sfgp:BaselineHazard]{sfgp::BaselineHazard}} -> \code{WeibullHazard}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-WeibullHazard-new}{\code{WeibullHazard$new()}}
\item \href{#method-WeibullHazard-clone}{\code{WeibullHazard$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="sfgp" data-topic="StanCodeCreator" data-id="print"><a href='../../sfgp/html/StanCodeCreator.html#method-StanCodeCreator-print'><code>sfgp::StanCodeCreator$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sfgp" data-topic="StanCodeCreator" data-id="stancode_data"><a href='../../sfgp/html/StanCodeCreator.html#method-StanCodeCreator-stancode_data'><code>sfgp::StanCodeCreator$stancode_data()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sfgp" data-topic="StanCodeCreator" data-id="stancode_functions"><a href='../../sfgp/html/StanCodeCreator.html#method-StanCodeCreator-stancode_functions'><code>sfgp::StanCodeCreator$stancode_functions()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sfgp" data-topic="StanCodeCreator" data-id="stancode_gq"><a href='../../sfgp/html/StanCodeCreator.html#method-StanCodeCreator-stancode_gq'><code>sfgp::StanCodeCreator$stancode_gq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sfgp" data-topic="StanCodeCreator" data-id="stancode_model"><a href='../../sfgp/html/StanCodeCreator.html#method-StanCodeCreator-stancode_model'><code>sfgp::StanCodeCreator$stancode_model()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sfgp" data-topic="StanCodeCreator" data-id="stancode_pars"><a href='../../sfgp/html/StanCodeCreator.html#method-StanCodeCreator-stancode_pars'><code>sfgp::StanCodeCreator$stancode_pars()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sfgp" data-topic="StanCodeCreator" data-id="stancode_tdata"><a href='../../sfgp/html/StanCodeCreator.html#method-StanCodeCreator-stancode_tdata'><code>sfgp::StanCodeCreator$stancode_tdata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sfgp" data-topic="StanCodeCreator" data-id="stancode_tpars"><a href='../../sfgp/html/StanCodeCreator.html#method-StanCodeCreator-stancode_tpars'><code>sfgp::StanCodeCreator$stancode_tpars()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sfgp" data-topic="StanCodeCreator" data-id="stanfiles_functions"><a href='../../sfgp/html/StanCodeCreator.html#method-StanCodeCreator-stanfiles_functions'><code>sfgp::StanCodeCreator$stanfiles_functions()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sfgp" data-topic="StanCodeCreator" data-id="string"><a href='../../sfgp/html/StanCodeCreator.html#method-StanCodeCreator-string'><code>sfgp::StanCodeCreator$string()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sfgp" data-topic="BaselineHazard" data-id="stanname_t"><a href='../../sfgp/html/BaselineHazard.html#method-BaselineHazard-stanname_t'><code>sfgp::BaselineHazard$stanname_t()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WeibullHazard-new"></a>}}
\if{latex}{\out{\hypertarget{method-WeibullHazard-new}{}}}
\subsection{Method \code{new()}}{
Create hazard
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WeibullHazard$new(
  prior_lambda = "gamma(5, 5)",
  prior_gamma = "inv_gamma(3, 6)"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{prior_lambda}}{prior of Weibull scale param}

\item{\code{prior_gamma}}{prior of Weibull shape param}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WeibullHazard-clone"></a>}}
\if{latex}{\out{\hypertarget{method-WeibullHazard-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WeibullHazard$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
