% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-sim.R
\name{sfsim}
\alias{sfsim}
\title{Simulate from SF model (3 arms)}
\usage{
sfsim(
  n1 = 4,
  n2 = 4,
  n3 = 4,
  ts0_scale = 20,
  mu_log_ks = c(3, 1.5, -1),
  mu_log_kg = c(-2, 0, 2),
  times = c(0.5, 3, 6, 12, 24, 48)/48,
  kg_sd = 0.05,
  ks_sd = 0.05,
  sigma = 0.2,
  x_effect = 0,
  log_C_kg = -2,
  log_C_ks = -1,
  t_sd = 0.02
)
}
\arguments{
\item{n1}{number of subjects in arm 1}

\item{n2}{number of subjects in arm 2}

\item{n3}{number of subjects in arm 3}

\item{ts0_scale}{initial tumor sizes will be sampled from the uniform
distribution \code{ts0_scale * U(1,3)}}

\item{mu_log_ks}{Mean log \code{k_s} parameters for the three arms.}

\item{mu_log_kg}{Mean log \code{k_g} parameters for the three arms.}

\item{times}{Time points where measurements are made. Scatter is added to
these.}

\item{kg_sd}{Jitter added to kg.}

\item{ks_sd}{Jitter added to ks.}

\item{sigma}{Noise param value.}

\item{x_effect}{Magnitude of effect of x on ks.}

\item{log_C_kg}{offset for log kg}

\item{log_C_ks}{offset for log ks}

\item{t_sd}{Jitter added to time points.}
}
\value{
a data frame
}
\description{
Simulate from SF model (3 arms)
}
