% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-treatment_effect.R
\name{treatment_effect}
\alias{treatment_effect}
\title{Analyse treatment effect}
\usage{
treatment_effect(
  fit_post,
  fit_prior,
  time_var = "t",
  t_pred = NULL,
  group_var = "arm",
  method = "new_sub"
)
}
\arguments{
\item{fit_post}{Posterior fit.}

\item{fit_prior}{Prior fit. Needs to have same number of draws as
\code{fit_post}.}

\item{time_var}{name of the time variable}

\item{t_pred}{Vector of time points (in days). Chosen based on data range
if \code{t_pred=NULL}. Chosen to be an evenly spaced vector from
0 to at most 2 years every 8 weeks if \code{t_pred="auto"}.}

\item{group_var}{grouping variable}

\item{method}{Used method for analyzing the treatment effect. Can be either
\itemize{
 \item \code{new_sub}: Simulating new imaginary subjects
 \item \code{group_est}: Using group-level parameter estimates
}}
}
\description{
Analyse treatment effect
}
